WIP implement project UI to show basic project metadata. See #1222.
