+++ src/ifcblenderexport/blenderbim/bim/__init__.py
@@ -46,7 +46,6 @@ if bpy is not None:
         operator.SelectDataDir,
         operator.SelectSchemaDir,
         operator.SelectIfcFile,
-        operator.ValidateIfcFile,
         operator.ExportIFC,
         operator.ImportIFC,
         operator.ColourByAttribute,
@@ -221,7 +220,6 @@ if bpy is not None:
         ui.BIM_PT_drawings,
         ui.BIM_PT_schedules,
         ui.BIM_PT_sheets,
-        ui.BIM_PT_bim,
         ui.BIM_PT_psets,
         ui.BIM_PT_classifications,
         ui.BIM_PT_document_information,
