+++ src/ifcblenderexport/blenderbim/bim/module/project/operator.py
@@ -1,4 +1,5 @@
 import bpy
+import logging
 import ifcopenshell
 import blenderbim.bim.module.project.create_file as create_file
 import blenderbim.bim.module.context.add_context as add_context
@@ -57,3 +58,16 @@ class CreateProject(bpy.types.Operator):
                 "relating_object": relating_object,
             },
         ).execute()
+
+
+class ValidateIfcFile(bpy.types.Operator):
+    bl_idname = "bim.validate_ifc_file"
+    bl_label = "Validate IFC File"
+
+    def execute(self, context):
+        import ifcopenshell.validate
+
+        logger = logging.getLogger("validate")
+        logger.setLevel(logging.DEBUG)
+        ifcopenshell.validate.validate(IfcStore.get_file(), logger)
+        return {"FINISHED"}
